javafx: adapt eaten for javafx and fix errors changed eaten to use bounds and made Food inherit from VisibleObject

+++ src/main/java/com/example/snakeproject/VisibleObject.java
@@ -9,8 +9,8 @@ public abstract class VisibleObject {
     double x;
     double y;
     Image i;
-    int w;
-    int h;
+    double w;
+    double h;
 
     public boolean l; //not necesary
